app/requirements.txt
# LangGraph & LangChain
langgraph>=0.2.0
langchain>=0.3.0
langchain-google-genai>=2.0.0
langchain-core>=0.3.0

# Google APIs
google-auth>=2.0.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.2.0
google-api-python-client>=2.0.0

# LINE Bot SDK
line-bot-sdk>=3.0.0

# Web Server
fastapi>=0.110.0
uvicorn[standard]>=0.29.0

# Scheduler
apscheduler>=3.10.0

# Utilities
python-dotenv>=1.0.0
httpx>=0.27.0
pydantic>=2.0.0
